added rest server for room info, update room table on refresh JOIN IS BROKEN
