+++ keys.go
@@ -5,10 +5,16 @@ import "github.com/charmbracelet/bubbles/key"
 type menuKeyMap struct {
 	Up    key.Binding
 	Down  key.Binding
-	Left  key.Binding
-	Right key.Binding
 	Quit  key.Binding
 	Enter key.Binding
+}
+
+type roomKeyMap struct {
+	Up      key.Binding
+	Down    key.Binding
+	Refresh key.Binding
+	Quit    key.Binding
+	Enter   key.Binding
 }
 
 type gameKeyMap struct {
@@ -17,7 +23,11 @@ type gameKeyMap struct {
 }
 
 func (k menuKeyMap) ShortHelp() []key.Binding {
-	return []key.Binding{k.Up, k.Down, k.Left, k.Right, k.Enter, k.Quit}
+	return []key.Binding{k.Up, k.Down, k.Enter, k.Quit}
+}
+
+func (k roomKeyMap) ShortHelp() []key.Binding {
+	return []key.Binding{k.Up, k.Down, k.Refresh, k.Enter, k.Quit}
 }
 
 func (k gameKeyMap) ShortHelp() []key.Binding {
@@ -25,6 +35,10 @@ func (k gameKeyMap) ShortHelp() []key.Binding {
 }
 
 func (k menuKeyMap) FullHelp() [][]key.Binding {
+	return [][]key.Binding{}
+}
+
+func (k roomKeyMap) FullHelp() [][]key.Binding {
 	return [][]key.Binding{}
 }
 
@@ -41,20 +55,35 @@ var menuKeys = menuKeyMap{
 		key.WithKeys("s", "down"),
 		key.WithHelp("↓/s", "move down"),
 	),
-	Left: key.NewBinding(
-		key.WithKeys("a", "left"),
-		key.WithHelp("←/a", "move left"),
+	Enter: key.NewBinding(
+		key.WithKeys("enter"),
+		key.WithHelp("enter", "make selection"),
 	),
-	Right: key.NewBinding(
-		key.WithKeys("d", "right"),
-		key.WithHelp("→/d", "move right"),
+	Quit: key.NewBinding(
+		key.WithKeys("q", "ctrl+c"),
+		key.WithHelp("q", "quit"),
+	),
+}
+
+var roomKeys = roomKeyMap{
+	Up: key.NewBinding(
+		key.WithKeys("w", "up"),
+		key.WithHelp("↑/w", "move up"),
+	),
+	Down: key.NewBinding(
+		key.WithKeys("s", "down"),
+		key.WithHelp("↓/s", "move down"),
+	),
+	Refresh: key.NewBinding(
+		key.WithKeys("r"),
+		key.WithHelp("r", "refresh rooms"),
 	),
 	Enter: key.NewBinding(
 		key.WithKeys("enter"),
 		key.WithHelp("enter", "make selection"),
 	),
 	Quit: key.NewBinding(
-		key.WithKeys("q", "esc", "ctrl+c"),
+		key.WithKeys("q", "ctrl+c"),
 		key.WithHelp("q", "quit"),
 	),
 }
@@ -65,7 +94,7 @@ var gameKeys = gameKeyMap{
 		key.WithHelp("1-9", "make move"),
 	),
 	Quit: key.NewBinding(
-		key.WithKeys("q", "esc", "ctrl+c"),
+		key.WithKeys("q", "ctrl+c"),
 		key.WithHelp("q", "quit"),
 	),
 }
